feat

+++ Assets/ScriptableObject/PlayerInfo.cs
@@ -12,7 +12,7 @@ public class PlayerInfo : ScriptableObject
     [Range(0f, 12f)]
     public int player_maxHp;
     [Range(0f, 12f)]
-    public int player_curHp;
+    public float player_curHp;
     public int curStage;
     public List<StageManager.MapConnetInfo> curStageInfoList;
     [HideInInspector]
